Runner could fail if screenshot timeout
Fix for the fail function in the runner. In the case of a misbehaving
website that timed out the screenshot function might also timeout as the
webdriver may be busy trying to retrieve the html page.
Fix for the fail function ensures that the entire test run is not broken
because of a single test/page fail.

Also made the screenshot save location configurable.