reporters: print all details of exception, not only the stack
It's fine if canopy wants to highlight the error by printing
the exception message first, the URL, and then the details.

But the details of it shouldn't be just the stacktrace because:
- It doesn't contain the type of the exception.
- It doesn't contain details about InnerExceptions.

I found this the hard way: receiving a "One or more errors
occurred" error, and not knowing what was going on for a while.
Later I realized that the exception was AggregateException
(which normally wraps exceptions that happen in async/parallel
code), and that it had the best explanation in the InnerException
of it.